renamed Operation->System and OperationInfo->Operation
